feat: parse relative and home directories in config files
